requirements.txt
streamlit==1.29.0
tensorflow==2.13.0
numpy
pillow